simplify package surface area, add docs, fix lint
Stop exporting the Parser type.  This was not thread safe, so a new one
need to be created for every parsed document anyway, and it didn't
expose any useful data.  Instead, move the Parse and ParseNode methods
to package funcs which build a new parser for each call.

Stop exporting a number of other symbols that almost certainly don't
need to be.  If they do prove to be useful, we can go back and export
them, but the focus here is to simplify things.

Additionally:
 - rename MicroFormat to Microformat
 - combine doc.go and microformat.go
 - add lots of godocs
 - fix misc lint (and similar) issues