move value parsing functions off parser type
These don't really need to be method of the parse type.  At most, a few
need access to the parser base URL, which we now pass it explicitly.
This should make these a little easier to test, and makes it clear that
they don't rely on the internal state of the parser.

Also cleanup a few minor things like variable names and using the zero
value bytes.Buffer.