add support for datetime values
this gets most of the way to fully supporting datetime values, though
there are still a few edge cases to add.  This still does not pass most
of the date related tests in microformats/tests, but that's mostly due
to (what I consider to be) issues with the tests themselves.